float changes made

+++ Week1&2/CLICalculator.py
@@ -1,10 +1,10 @@
 while True:
-    num1 = int(input("Enter the first number: "))
+    num1 = float(input("Enter the first number: "))
     num2 = ""
     operator = ""
     
     if (num1 / 1 == num1):
-        num2 = int(input("Enter the second number: "))
+        num2 = float(input("Enter the second number: "))
     else:
         print("Invalid input")
     if (num2 / 1 == num2):
